+++ include/switches.hpp
@@ -4,18 +4,18 @@
 class SWITCHES
 {
     public:
-    void begin(void (*fire_event)(uint32_t event, uint8_t event_subtype, event_data ed), uint8_t ten_ms_time_slot = 0){
+    void begin(void (*fire_event)(uint32_t event, uint32_t event_subtype, event_data ed), uint8_t ten_ms_time_slot = 0){
         event_fire = fire_event;
         time_slot = ten_ms_time_slot;
     }
 
    
-    void handler(event_data ed, uint8_t event_subtype);
+    void handler(event_data ed, uint32_t event_subtype);
    
     private:
     uint8_t time_slot;
     uint8_t ms_counter;
-    void (*event_fire)(uint32_t event, uint8_t event_subtype, event_data ed);
+    void (*event_fire)(uint32_t event, uint32_t event_subtype, event_data ed);
 
 };
